Refactor to handle collision logic in each actor instead of big switch statement in GameplayState
Break out the actor collision logic from the GameplayState switch case statement such that each actor type is now responsible for what happens when it collides with the player.

Additionally, inventory items are now a separate class from placeable actors, with keys represented as one class when on the maze, and another class once its in inventory.