Merge client and server into separate classes in one project
Also, implement a NetworkMessage class for all messages to inherit from, to distinguish between messages to set your own username from text messages.

This still needs a lot of work on the UI/UX side, but in terms of sending messages back and forth and using a multithreaded client, I think it works.